+++ dts/save-structure/const-data/index.d.ts
@@ -1,4 +1,5 @@
 export * from "./accessories";
 export * from "./geysers";
 export * from "./skills";
+export * from "./space-destinations";
 export * from "./template-enumerations";