+++ src/save-structure/game-objects/game-object-behavior/known-behaviors/spacecraft-manager.ts
@@ -1,6 +1,7 @@
-import { SimHashes } from "../../../const-data";
+import { SimHashes, SpaceDestinationTypeName } from "../../../const-data";
 
 import { BehaviorName } from "./types";
+import { MinionResumeBehavior } from "./minion-resume";
 
 export const SpacecraftManagerBehavior: BehaviorName<
   SpacecraftManagerBehavior
@@ -10,51 +11,55 @@ export interface SpacecraftManagerBehavior {
   name: "SpacecraftManager";
   templateData: {
     spacecraft: any[];
+    analyzeDestinationID: number;
+    destinationAnalysisScores: ([number, number])[];
     destinations: SpaceDestination[];
+    savedSpacecraftDestinations: ([number, number])[];
+    nextSpacecraftID: number;
+    destinationsGenerated: boolean;
   };
 }
-export interface SpaceDestination {
+
+export interface Spacecraft {
   id: number;
-  type: SpaceDestinationType;
+  rocketName: string;
+  refLaunchConditions: any;
+  moduleCount: number;
+  missionState: MissionState;
+  missionElapsed: number;
+  missionDuration: number;
+}
+
+export enum MissionState {
+  Grounded = "Grounded",
+  Launching = "Launching",
+  Underway = "Underway",
+  WaitingToLand = "WaitingToLand",
+  Landing = "Landing",
+  Destroyed = "Destroyed"
+}
+
+export interface SpaceDestination {
+  id: MinionResumeBehavior;
+  type: SpaceDestinationTypeName;
   distance: number;
+  startingOrbitPercentage: number;
+  availableMass: number;
   activePeriod: number;
   inactivePeriod: number;
-  startingOrbitPercentage: number;
+  /**
+   * All elements recoverable by missions to this destination.
+   * An array of tuples.  First tuple item is the sim hash
+   * of the element, the second tuple item is a value 0 to 1 representing
+   * the random factor for how much of this element to recover.
+   */
   recoverableElements: ([SimHashes, number])[];
-  researchOpportunities: SpacecraftResearchOpportunity;
+  researchOpportunities: SpacecraftResearchOpportunity[];
 }
 export interface SpacecraftResearchOpportunity {
-  description: number;
+  description: string;
   dataValue: number;
   completed: boolean;
   discoveredRareResource: SimHashes;
   discoveredRareItem: string;
 }
-
-export enum SpaceDestinationType {
-  Satellite = "Satellite",
-  MetallicAsteroid = "MetallicAsteroid",
-  RockyAsteroid = "RockyAsteroid",
-  CarbonaceousAsteroid = "CarbonaceousAsteroid",
-  IcyDwarf = "IcyDwarf",
-  OrganicDwarf = "OrganicDwarf",
-  DustyMoon = "DustyMoon",
-  TerraPlanet = "TerraPlanet",
-  VolcanoPlanet = "VolcanoPlanet",
-  GasGiant = "GasGiant",
-  IceGiant = "IceGiant"
-}
-
-export const SpaceDestinationTypes: SpaceDestinationType[] = [
-  SpaceDestinationType.Satellite,
-  SpaceDestinationType.MetallicAsteroid,
-  SpaceDestinationType.RockyAsteroid,
-  SpaceDestinationType.CarbonaceousAsteroid,
-  SpaceDestinationType.IcyDwarf,
-  SpaceDestinationType.OrganicDwarf,
-  SpaceDestinationType.DustyMoon,
-  SpaceDestinationType.TerraPlanet,
-  SpaceDestinationType.VolcanoPlanet,
-  SpaceDestinationType.GasGiant,
-  SpaceDestinationType.IceGiant
-];